+++ doc/source/doxygen-docs/changelog.md
@@ -18,6 +18,8 @@
     - mrpt::hwdrivers::CPtuDPerception
     - mrpt::hwdrivers::CRoboticHeadInterface
     - mrpt::hwdrivers::CTuMicos
+- BUG FIXES:
+  - mrpt::hwdrivers::CJoystick: Crash due to missing #include of OS-specific configuration in CJoystick.h
 
 # Version 2.12.2: Released April 23rd, 2024
 - Changes in libraries: